refactor: Déplace composants ui dans dossier ui
