+++ components/jeune/FormulaireRechercheDossier.tsx
@@ -1,5 +1,5 @@
-import Button from 'components/Button'
-import { ErrorMessage } from 'components/ErrorMessage'
+import Button from 'components/ui/Button'
+import { ErrorMessage } from 'components/ui/ErrorMessage'
 import { useRouter } from 'next/router'
 import { FormEvent, useEffect, useState } from 'react'
 import IndicationRechercheDossier from 'components/jeune/IndicationRechercheDossier'
